backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
Django==4.2.11
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
setuptools==69.2.0
soupsieve==2.5
sqlparse==0.4.4
uritemplate==4.1.1
